T-017: PR-3: Create prop-driven ai_summary_section.tsx in v2 + update legacy
